+++ __tests__/support/faker.js
@@ -1,0 +1,7 @@
+/* global jest */
+
+const { faker } = require('@faker-js/faker/locale/en')
+
+faker.seed(jest.getSeed())
+
+module.exports = faker