+++ spec/dummy/app/models/list.rb
@@ -2,7 +2,11 @@
 
 class List < ApplicationRecord
   # Relations
-  has_many :items, -> { order('position') }, dependent: :destroy
+  has_many :items,
+           -> { order('position') },
+           inverse_of: :list,
+           dependent: :destroy
+
   accepts_nested_attributes_for :items, reject_if: :all_blank, allow_destroy: true
 
   # Validations