+++ spec/dummy/config/initializers/assets.rb
@@ -6,11 +6,9 @@
 Rails.application.config.assets.version = '1.0'
 
 # Add additional assets to the asset load path.
-require 'jasmine-core'
-Rails.application.config.assets.paths << File.dirname(Jasmine::Core.path)
-Rails.application.config.assets.paths << Rails.root.join('../javascripts')
+# Rails.application.config.assets.paths << Rails.root.join('../dir')
 
 # Precompile additional assets.
 # application.js, application.css, and all non-JS/CSS in the app/assets
 # folder are already added.
-Rails.application.config.assets.precompile += %w[jasmine.js jasmine.css]
+# Rails.application.config.assets.precompile += %w[file.js file.css]